+++ lib/src/painter.dart
@@ -1,3 +1,4 @@
+import 'package:countries_world_map/src/helpers/map_instructions.dart';
 import 'package:flutter/material.dart';
 import '../components/canvas/touchy_canvas.dart';
 import '../countries_world_map.dart';
@@ -95,42 +96,4 @@ class SimpleMapPainter extends CustomPainter {
   @override
   bool shouldRepaint(SimpleMapPainter oldDelegate) =>
       oldDelegate.colors != colors;
-}
-
-class SimpleMapInstruction {
-  /// uniqueID of the territory being drawn
-  String uniqueID;
-
-  /// Name of the territory being drawn
-  String name;
-
-  /// List of instructions to draw the territory
-  List<String> instructions;
-
-  SimpleMapInstruction(
-      {required this.uniqueID, required this.instructions, required this.name});
-
-  // To Json
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{
-      "\"n\"": "\"$name\"",
-      "\"u\"": "\"$uniqueID\"",
-      "\"i\"": instructions,
-    };
-    return data;
-  }
-
-  // From Json
-  factory SimpleMapInstruction.fromJson(Map<String, dynamic> json) {
-    List<String> paths = <String>[];
-
-    List jsonPaths = json['i'];
-
-    for (int i = 0; i < jsonPaths.length; i++) {
-      paths.add(jsonPaths[i]);
-    }
-
-    return SimpleMapInstruction(
-        uniqueID: json['u'], name: json['n'], instructions: paths);
-  }
 }